Método listarMinhasInscricoes movido para MeusInscricoesAdapter
